# backend/openapi.yaml
openapi: 3.0.3
info:
  title: Kavach Parental Control API
  description: API for the Kavach parental control application
  version: 1.0.0
servers:
  - url: http://localhost:3000/api/v1
security:
  - bearerAuth: []
  - sessionCookie: []
...
paths:
...
  /children:
    get:
      tags: [Children]
      summary: List children
      operationId: listChildren
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: Paginated list of children
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: "#/components/schemas/Child"
                      pagination:
                        $ref: "#/components/schemas/Pagination"
    post:
      tags: [Children]
      summary: Add a child
      operationId: createChild
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
                birth_date:
                  type: string
                  format: date
      responses:
        "201":
          description: Child created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Envelope"
...
  /keywords:
    get:
      tags: [Keyword Dictionary]
      summary: List keywords in the detection dictionary
      operationId: listKeywords
      parameters:
        - name: category
          in: query
          schema:
            type: string
            enum: [CYBERBULLYING, SELF_HARM, PROFANITY, DRUGS, CUSTOM]
        - name: active_only
          in: query
          schema:
            type: boolean
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: List of keywords
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: "#/components/schemas/KeywordEntry"
                      pagination:
                        $ref: "#/components/schemas/Pagination"
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
    post:
      tags: [Keyword Dictionary]
      summary: Add a keyword to the dictionary
      operationId: createKeyword
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateKeywordInput"
      responses:
        "201":
          description: Keyword created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Envelope"
        "422":
          description: Validation error
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
...
  /integrations:
    get:
      tags: [Integrations]
      summary: List all integrations
      operationId: listIntegrations
      responses:
        "200":
          description: List of integrations
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: "#/components/schemas/Integration"
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
    post:
      tags: [Integrations]
      summary: Create a new integration
      operationId: createIntegration
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateIntegrationInput"
      responses:
        "201":
          description: Integration created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Envelope"
        "422":
          description: Validation error
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    sessionCookie:
      type: apiKey
      in: cookie
      name: kavach_access
  parameters:
    Page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
...
  schemas:
    Envelope:
      type: object
      properties:
        success:
          type: boolean
        data:
          oneOf:
            - type: object
            - type: array
            - type: string
            - type: number
            - type: boolean
            - type: "null"
          nullable: true
        error:
          type: string
          nullable: true
        timestamp:
          type: string
          format: date-time
        request_id:
          type: string
          format: uuid
    Pagination:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        pages:
          type: integer
...
    Child:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        birth_date:
          type: string
          format: date
          nullable: true
...
    KeywordEntry:
      type: object
      properties:
        id:
          type: string
          format: uuid
        category:
          type: string
          enum: [CYBERBULLYING, SELF_HARM, PROFANITY, DRUGS, CUSTOM]
        keyword:
          type: string
        severity:
          type: string
          enum: [LOW, MEDIUM, HIGH, CRITICAL]
        language:
          type: string
        is_active:
          type: boolean
        created_at:
          type: string
          format: date-time
    CreateKeywordInput:
      type: object
      required: [category, keyword]
      properties:
        category:
          type: string
          enum: [CYBERBULLYING, SELF_HARM, PROFANITY, DRUGS, CUSTOM]
        keyword:
          type: string
          minLength: 1
          maxLength: 255
        severity:
          type: string
          enum: [LOW, MEDIUM, HIGH, CRITICAL]
          default: MEDIUM
        language:
          type: string
          maxLength: 10
          default: en
        is_active:
          type: boolean
          default: true
...
    Integration:
      type: object
      properties:
        id:
          type: string
          format: uuid
        integration_type:
          type: string
          enum: [SCHOOL_PORTAL, CALENDAR, HEALTH_APP, CUSTOM]
        name:
          type: string
        config:
          type: object
        is_active:
          type: boolean
        last_synced_at:
          type: string
          format: date-time
          nullable: true
        created_at:
          type: string
          format: date-time
    CreateIntegrationInput:
      type: object
      required: [integration_type, name]
      properties:
        integration_type:
          type: string
          enum: [SCHOOL_PORTAL, CALENDAR, HEALTH_APP, CUSTOM]
        name:
          type: string
          minLength: 1
          maxLength: 255
        config:
          type: object
...